# Encoding speedups for strings and bools: