Some speedups on string and boolean encoding primarily by reducing the amount of allocations required, thus reducing gc garbage as well.  The string encoding speedup was pretty significant, because it saves an extra copy of the whole string, so the longer the string the better the speedup, but even on short strings you get 21.66% speedup.

String results:

```
$ benchcmp a.txt b.txt 
benchmark                              old ns/op     new ns/op     delta
BenchmarkStringSerialization_small     157           123           -21.66%
BenchmarkStringSerialization_med       181           126           -30.39%
BenchmarkStringSerialization_large     730           186           -74.52%
```

Boolean results:

```
$ benchcmp a.txt b.txt 
benchmark                           old ns/op     new ns/op     delta
BenchmarkBooleanSerialization-4     36.3          21.8          -39.94%
```